ДЗ-9 промежуточный коммит все методы отлажены и работают, но нужно еще написать тесты

+++ hw09-jdbc/src/main/java/ru/otus/api/model/User.java
@@ -19,7 +19,7 @@ public class User {
     @Size(size = 3)
     private int age;
 
-    public User(int id, String name, int age) {
+    public User(long id, String name, int age) {
         this.id = id;
         this.name = name;
         this.age = age;
